+++ NEWS.md
@@ -1,3 +1,5 @@
+# dymiumCore (development version)
+
 # dymiumCore 0.1.8
 
 - Added `sim()` for compiling and executing a microsimulation pipeline.